+++ src/ExceptionDataBag.php
@@ -13,7 +13,7 @@ namespace Sentry;
 final class ExceptionDataBag
 {
     /**
-     * @var class-string<\Throwable> The type of exception, e.g. RuntimeException
+     * @var string The type of exception, e.g. RuntimeException
      */
     private $type;
 
@@ -42,12 +42,20 @@ final class ExceptionDataBag
 
     /**
      * Gets the type of exception, e.g. RuntimeException.
-     *
-     * @return class-string<\Throwable>
      */
     public function getType(): string
     {
         return $this->type;
+    }
+
+    /**
+     * Sets the type of the exception.
+     *
+     * @param string $type The exception type
+     */
+    public function setType(string $type): void
+    {
+        $this->type = $type;
     }
 
     /**